apps/hip-3-pusher/pyproject.toml
[project]
name = "hip-3-pusher"
version = "0.2.7"
description = "Hyperliquid HIP-3 market oracle pusher"
readme = "README.md"
requires-python = "==3.13.*"
dependencies = [
    "boto3~=1.40.38",
    "cryptography~=46.0.1",
    "httpx~=0.28.1",
    "hyperliquid-python-sdk~=0.19.0",
    "loguru~=0.7.3",
    "opentelemetry-exporter-prometheus~=0.58b0",
    "opentelemetry-sdk~=1.37.0",
    "prometheus-client~=0.23.1",
    "tenacity~=9.1.2",
    "websockets~=15.0.1",
]

[build-system]
requires = ["setuptools~=80.9", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = ["pytest~=8.4.2", "mypy~=1.15.0", "boto3-stubs[kms]~=1.40.38"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
ignore_missing_imports = true
exclude = ["tests/", "scripts/", "config/"]
